Follow MW conventions regarding lines continuation
https://www.mediawiki.org/wiki/Manual:Coding_conventions#Line_continuation

The suggestion is about moving the logical operators in conditions to the beginning of a line and keeping the first check in the same line as a command. I personally find it helpful when I grep our codebase or use the Refactor mechanism in PHP storm. If a piece of a code is formatted this way every line has a clear indicator of being part of a logical condition.